redo how to create dummy columns in bis init

+++ sql/now/_init.sql
@@ -193,13 +193,13 @@ JOBNUMBER_relevant as (
 	NULL::text as zsf_prop,
 	-- Requested enhancement from Housing with new columns from DOB source data
 	(CASE WHEN uselabel ~* 'Residential' THEN total_floor_area 
-		ELSE NULL END) as ZSF_R_prop,
+		ELSE NULL END)::numeric as ZSF_R_prop,
 	(CASE WHEN uselabel ~* 'Commercial' THEN total_floor_area 
-		ELSE NULL END) as ZSF_C_prop,
+		ELSE NULL END)::numeric as ZSF_C_prop,
 	(CASE WHEN uselabel ~* 'Community Facility' THEN total_floor_area 
-		ELSE NULL END) as ZSF_CF_prop,
+		ELSE NULL END)::numeric as ZSF_CF_prop,
 	(CASE WHEN uselabel ~* 'Manufacturing' THEN total_floor_area 
-		ELSE NULL END) as ZSF_M_prop,
+		ELSE NULL END)::numeric as ZSF_M_prop,
 	no_of_parking_spaces::numeric as Prkng_prop,
 	building_type as bldg_class,
 	NULL as desc_other,
